requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
email-validator==2.1.1
bcrypt==4.2.0
passlib==1.7.4
sqladmin==0.22.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
itsdangerous>=2.0.0
starlette>=0.38.0
werkzeug>=3.0.0
PyJWT==2.10.1
